Move variables related to encrypted exports into base component.
